requirements.txt
requests>=2.31.0
pandas>=2.0.0
geopy>=2.4.0
streamlit>=1.30.0
folium>=0.14.0
streamlit-folium>=0.15.0
python-dotenv==1.0.1
tabulate==0.9.0
together==0.2.5
typing-extensions>=4.7.0 